fix/ loginBtnChangeAction,editPageActiveAll

+++ src/pages/EditProfile.jsx
@@ -5,6 +5,8 @@ import supabase from '../supabase/supabase';
 import { v4 as uuidv4 } from 'uuid';
 import useFormInputs from '../hooks/useInput';
 import { checkSignIn } from '../redux/slices/authSlice';
+import { logout } from '../redux/slices/authSlice';
+import { validatePasswordFormat, validatePasswordMatch } from '../utils/validators';
 
 const initialState = {
   name: '',
@@ -18,10 +20,10 @@ function EditProfile() {
   const dispatch = useDispatch();
   const navigate = useNavigate();
   const user = useSelector((state) => state.auth.user);
-  const { inputs, setInputs, handleOnChange, handleResetInputs } = useFormInputs(initialState);
+  const { inputs, handleOnChange} = useFormInputs(initialState);
   const { name, nickname, currentPassword, mylogReason, profilePicture } = inputs;
   const [picture, setPicture] = useState('');
-
+  const [errors, setErrors] = useState({});
   useEffect(() => {
     dispatch(checkSignIn());
   }, [dispatch]);
@@ -58,26 +60,43 @@ function EditProfile() {
       inputs.profilePicture = profilePictureUrl.publicUrl;
     }
   };
-  //supabase 로직을 사용(update 또는 edit)하여 변경사항 가능하게 적용
   const handleChangeProfile = async () => {
-    console.log(currentPassword);
-    console.log(user.id);
-    console.log(picture);
-    const { data, error } = await supabase
-      .from('Users')
-      .update({
-        name: name,
-        nickname: nickname,
-        password: currentPassword,
-        reason: mylogReason,
-        imageUrl: picture
-      })
-      .eq('userId', user.id);
+    const newErrors = {};
 
-    if (error) {
-      console.error('업데이트에 문제가 발생했습니다.', error);
-    } else {
-      console.log('업데이트가 완료 됐습니다!', data);
+    try {
+      if (!name || !currentPassword) {
+        newErrors.general = '모든 필드를 입력해주세요.';
+        throw new Error('모든 필드를 입력해주세요.');
+      }
+      if (!validatePasswordFormat(currentPassword)) {
+        newErrors.password = '비밀번호는 영문 대소문자, 특수문자를 포함하여 8자리 이상이어야 합니다.';
+        throw new Error('비밀번호는 영문 대소문자, 특수문자를 포함하여 8자리 이상이어야 합니다.');
+      }
+
+      if (!validatePasswordMatch(currentPassword, data.password)) {
+        newErrors.unPassword = '비밀번호가 일치하지 않습니다.';
+        throw new Error('비밀번호 불일치');
+      }
+
+      const { data, error } = await supabase
+        .from('Users')
+        .update({
+          name: name,
+          nickname: nickname,
+          password: currentPassword,
+          reason: mylogReason,
+          imageUrl: picture
+        })
+        .eq('userId', user.id);
+
+      if (error) {
+        console.error('업데이트에 문제가 발생했습니다.', error);
+      } else {
+        return alert('업데이트가 완료 됐습니다!', data);
+      }
+    } catch (err) {
+      setErrors(newErrors);
+      console.log(err);
     }
   };
 
@@ -88,19 +107,15 @@ function EditProfile() {
 
   // 탈퇴하기
   const handleDeleteAccount = async () => {
-    const { error } = await supabase
-    .from('Users')
-    .delete()
-    .eq('userId', user.id);
-
+    const { error } = await supabase.from('Users').delete().eq('userId', user.id);
     if (error) {
       alert('계정 삭제 중 오류가 발생했습니다' + error.message);
     } else {
       const { error } = await supabase.auth.signOut();
-
       if (error) {
-        alert('로그아웃 중 오류가 발생했습니다' );
+        alert('로그아웃 중 오류가 발생했습니다');
       } else {
+        dispatch(logout());
         alert('계정이 성공적으로 삭제되었습니다.');
         navigate('/');
       }
@@ -136,12 +151,15 @@ function EditProfile() {
             value={currentPassword}
             onChange={handleOnChange}
           />
+          {errors.password && <div style={{ color: 'red' }}>{errors.password}</div>}
+          {errors.unPassword && <div style={{ color: 'red' }}>{errors.unPassword}</div>}
         </div>
 
         <div>
           <label htmlFor="reasonToUseBlog">Mylog 운영 이유:</label>
           <input type="text" id="mylogReason" name="mylogReason" value={mylogReason} onChange={handleOnChange} />
         </div>
+        {errors.general && <div style={{ color: 'red' }}>{errors.general}</div>}
         <button type="submit" style={{ color: 'blue' }} onClick={handleChangeProfile}>
           변경 완료
         </button>
